Fixed remaining String constructors
Added nativeaot version of MemoryMarshal to handle array reference.
