+++ EFISharp.CoreLib/System/Runtime/CompilerServices/RuntimeHelpers.CoreRT.cs
@@ -62,6 +62,24 @@ namespace System.Runtime.CompilerServices
 
     }
 
+    // CLR arrays are laid out in memory as follows (multidimensional array bounds are optional):
+    // [ sync block || pMethodTable || num components || MD array bounds || array data .. ]
+    //                 ^               ^                 ^                  ^ returned reference
+    //                 |               |                 \-- ref Unsafe.As<RawArrayData>(array).Data
+    //                 \-- array       \-- ref Unsafe.As<RawData>(array).Data
+    // The BaseSize of an array includes all the fields before the array data,
+    // including the sync block and method table. The reference to RawData.Data
+    // points at the number of components, skipping over these two pointer-sized fields.
+    [StructLayout(LayoutKind.Sequential)]
+    internal class RawArrayData
+    {
+        public uint Length; // Array._numComponents padded to IntPtr
+#if TARGET_64BIT
+        public uint Padding;
+#endif
+        public byte Data;
+    }
+
     [StructLayout(LayoutKind.Sequential)]
     internal class RawData
     {
